Make all checkbox spinning to make it more obvious to the user
